Move namespace-aliasing code into separate class
