+++ MonoGame.Framework.Content.Pipeline/Serialization/Intermediate/IntermediateSerializer.cs
@@ -56,6 +56,7 @@ namespace Microsoft.Xna.Framework.Content.Pipeline.Serialization.Intermediate
         private IntermediateSerializer()
         {
             _scannedObjects = new List<object>();
+            _namespaceAliasHelper = new NamespaceAliasHelper(this);
         }
 
         /// <summary>
@@ -63,20 +64,11 @@ namespace Microsoft.Xna.Framework.Content.Pipeline.Serialization.Intermediate
         /// </summary>
         private Dictionary<string, string> _namespaceLookup;
 
-        /// <summary>
-        /// Maps "My.Namespace.LongName" -> "ShortName" for type lookups.
-        /// </summary>
-        private Dictionary<string, AliasedNamespace> _namespaceLookupReverse;
-
-        private class AliasedNamespace
-        {
-            public string Alias;
-            public string TypePrefix;
-        }
-
         private Dictionary<Type, ContentTypeSerializer> _serializers;
 
         private Dictionary<Type, Type> _genericSerializerTypes;
+
+        private readonly NamespaceAliasHelper _namespaceAliasHelper;
 
         private readonly List<object> _scannedObjects;
 
@@ -194,7 +186,7 @@ namespace Microsoft.Xna.Framework.Content.Pipeline.Serialization.Intermediate
             var writer = new IntermediateWriter(serializer, output, referenceRelocationPath);
             output.WriteStartElement("XnaContent");
 
-            serializer.WriteNamespaces(output, value);
+            serializer._namespaceAliasHelper.WriteNamespaces(output, value);
 
             // Write the asset.
             var format = new ContentSerializerAttribute { ElementName = "Asset" };
@@ -224,132 +216,6 @@ namespace Microsoft.Xna.Framework.Content.Pipeline.Serialization.Intermediate
 
                 _namespaceLookup.Add(reader.LocalName + ":", reader.Value + ".");
             }
-        }
-
-        private void WriteNamespaces<T>(XmlWriter writer, T value)
-        {
-            // Maps "My.Namespace.LongName" -> "ShortName" for type lookups.
-            _namespaceLookupReverse = new Dictionary<string, AliasedNamespace>();
-
-            var childNamespaces = new List<string>();
-            ContentTypeSerializer.ChildCallback onScanChild = (contentTypeSerializer, child) =>
-            {
-                if (child == null)
-                    return;
-
-                var childType = child.GetType();
-
-                if (contentTypeSerializer.TargetType == childType)
-                    return;
-
-                if (_typeAliasesReverse.ContainsKey(childType))
-                    return;
-
-                var childNamespace = childType.Namespace;
-
-                if (string.IsNullOrEmpty(childNamespace))
-                    return;
-
-                childNamespaces.Add(childNamespace);
-            };
-
-            // Force top-level object type to be included.
-            onScanChild(GetTypeSerializer(typeof(object)), value);
-
-            var serializer = GetTypeSerializer(typeof(T));
-            serializer.ScanChildren(this, onScanChild, value);
-
-            childNamespaces = childNamespaces.Distinct().ToList();
-
-            // Do first pass to determinate what our aliases are.
-            var sortedChildNamespaces = new List<string>(childNamespaces);
-            sortedChildNamespaces.Sort();
-            var tempAliases = new Dictionary<string, string>();
-            foreach (var childNamespace in sortedChildNamespaces)
-            {
-                var alias = FindAlias(tempAliases, childNamespace);
-                if (alias != null)
-                    tempAliases.Add(childNamespace, alias);
-            }
-
-            // Do second pass to calculate the TypePrefix for each alias.
-            foreach (var childNamespace in childNamespaces)
-            {
-                var alias = FindAvailableAlias(tempAliases, childNamespace, childNamespace);
-                if (alias != null)
-                    _namespaceLookupReverse.Add(childNamespace, alias);
-            }
-
-            var writtenAliases = new List<string>();
-            foreach (var kvp in _namespaceLookupReverse)
-            {
-                if (!string.IsNullOrEmpty(kvp.Value.TypePrefix))
-                    continue;
-                if (!writtenAliases.Contains(kvp.Value.Alias))
-                    writer.WriteAttributeString("xmlns", kvp.Value.Alias, null, kvp.Key);
-                writtenAliases.Add(kvp.Value.Alias);
-            }
-        }
-
-        private string FindAlias(Dictionary<string, string> aliases, string childNamespace)
-        {
-            if (string.IsNullOrEmpty(childNamespace))
-                return null;
-
-            var alias = childNamespace.Substring(childNamespace.LastIndexOf('.') + 1);
-            if (aliases.All(x => x.Value != alias))
-                return alias;
-
-            // Find the longest parent namespace.
-            if (aliases.Any(x => childNamespace.StartsWith(x.Key)))
-            {
-                string longestParentNamespace = string.Empty;
-                foreach (var kvp in aliases)
-                {
-                    if (childNamespace.StartsWith(kvp.Key) && kvp.Key.Length > longestParentNamespace.Length)
-                        longestParentNamespace = kvp.Key;
-                }
-                return aliases[longestParentNamespace];
-            }
-
-            return null;
-        }
-
-        private AliasedNamespace FindAvailableAlias(Dictionary<string, string> tempAliases, string childNamespace, string originalNamespace)
-        {
-            string alias;
-            if (tempAliases.TryGetValue(childNamespace, out alias))
-            {
-                if (alias == childNamespace.Substring(childNamespace.LastIndexOf('.') + 1))
-                    return new AliasedNamespace
-                    {
-                        Alias = alias,
-                        TypePrefix = string.Empty
-                    };
-
-                var namespaceParent = tempAliases
-                    .Where(x => x.Value == alias)
-                    .Select(x => x.Key)
-                    .OrderBy(x => x.Length)
-                    .First();
-                return new AliasedNamespace
-                {
-                    Alias = alias,
-                    TypePrefix = GetRelativeNamespace(namespaceParent, childNamespace) + "."
-                };
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// Returns just the portion <paramref name="@namespace"/> relative to <paramref name="namespaceParent"/>.
-        /// For example, given namespaceParent=Foo.Bar and @namespace=Foo.Bar.Baz, will return Baz.
-        /// </summary>
-        private static string GetRelativeNamespace(string namespaceParent, string @namespace)
-        {
-            if (@namespace.StartsWith(namespaceParent))
-                return @namespace.Substring(namespaceParent.Length + 1);
-            throw new InvalidOperationException();
         }
 
         /// <summary>
@@ -399,13 +265,10 @@ namespace Microsoft.Xna.Framework.Content.Pipeline.Serialization.Intermediate
                 return typeName;
 
             // Look for aliased namespace.
-            if (!string.IsNullOrEmpty(type.Namespace))
-            {
-                AliasedNamespace namespaceAlias;
-                if (_namespaceLookupReverse.TryGetValue(type.Namespace, out namespaceAlias))
-                    return namespaceAlias.Alias + ":" + namespaceAlias.TypePrefix + type.Name;
-            }
+            if (_namespaceAliasHelper.TryGetAliasedTypeName(type, out typeName))
+                return typeName;
 
+            // Fallback to full type name.
             return type.FullName;
         }
 
@@ -415,6 +278,11 @@ namespace Microsoft.Xna.Framework.Content.Pipeline.Serialization.Intermediate
                 return true;
             _scannedObjects.Add(value);
             return false;
+        }
+
+        internal bool HasTypeAlias(Type type)
+        {
+            return _typeAliasesReverse.ContainsKey(type);
         }
     }
 }
